Refactor: refactored files to be 1 file per plugin
